+++ backend/src/modules/AddComment/CommentModule.js
@@ -1,0 +1,19 @@
+import { CommentController } from './controller/Comment.controller.js';
+import { CommentService } from './service/Comment.service.js';
+import { InMemoryCommentRepository } from './repository/inMemory/inMemoryComment.repository.js';
+import { MongoCommentRepository } from './repository/mongo/MongoComment.repository.js';
+
+export const commentModule = (moduleRepositoryType) => {
+  const addCommentRepository = (repositoryType) => {
+    if (repositoryType === 'MONGO') {
+      return new MongoCommentRepository();
+    }
+    return new InMemoryCommentRepository();
+  };
+
+  const repository = addCommentRepository(moduleRepositoryType);
+  const service = new CommentService(repository);
+  const controller = new CommentController(service);
+
+  return controller.router;
+};
